# docs/swagger.yaml
basePath: /
definitions:
  models.SignIn:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.SignUp:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      is_registered:
        type: boolean
      name:
        type: string
      pass_hash:
        type: string
      tg_chat_id:
        type: integer
      who:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для бронирования коворкингов
  title: CoworkGo API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя
      parameters:
      - description: SignIn Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя
      parameters:
      - description: Данные регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SignUp'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
swagger: "2.0"